Pulled sub commands out to separate files.
Keeping the simulation and the post-processing commands in the same file
was nice because it was so easy to share functions. But the problem with
that approach was: changes to the plot routines also forces me to re-run
the simulations. The depdenency graph will be better if I keep those
files separate. I have refactored to make it so.

At this point, simulate() and batch() are still in the same file. I like
that _simulate() is still an internal function. Changes to the
simulation or the batching both require me to re-run the simulations
anyway, so this should be fine.