Refactored to appease pycodestyle.
I'm not sure if this improved or deteriorated readability in general.
I like the changes I made to the analysis functions, but I'm not sure
about renaming the variables in `_simulate()`. I think it was easier to
read "R" as recovered rather than "recovered". Lots of visual noise that
looks less like the math that the code solves. Also, I switched to
storing values directly in the h5 file instead of generating a list and
then writing all at once. Some upsides to that - if a sim were to crash
half way though, you could look at some evidence of what happened. Also,
I haven't measured, but I think h5py is buffering writes anyway, so I
probably saved some memory and didn't increase time. Doesn't matter for
a tiny toy simulation like this, but might be relevant in a larger one.
And we're just studying techniques here.